pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "config-validator"
version = "0.2.0"
description = "Advanced configuration validation library"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "PyYAML>=6.0",
    "toml>=0.10.2"
]

[project.scripts]
validate-config = "config_validator.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "mypy>=1.0"]